Remove functions to get config sub values.
All the actual values of the config need to be accessed via functions to allow for logic in the future however I don't think accessing the structs which hold these values via a function is needed, the only benefit I see is that I can enforce using a reference. Hope that I never partially move it whoops.